+++ lib/WriteOff/Controller/Vote/Public.pm
@@ -16,30 +16,29 @@ Catalyst Controller.
 
 =cut
 
+sub init :Private {
+	my ( $self, $c ) = @_;
+	
+	$c->forward('/captcha_get');
+	
+	$c->stash->{formid} = join "|", 
+		'form', 'event', $c->stash->{event}->id, 'public', $c->action->name;
+		
+	push $c->stash->{title}, 'Vote', 'Public';
+}
+
 sub fic :PathPart('vote/public') :Chained('/event/fic') :Args(0) {
     my ( $self, $c ) = @_;
 
-	$c->forward('/captcha_get');
+	$c->forward('init');
 	
-	my $formid = "form" . "event" . $c->stash->{event}->id . "public" . "fic";
-	
-	if( $c->req->method eq 'POST' ) {
-		if( $c->req->params->{submit} eq 'Save vote' ) {
-			$c->session->{$formid} = $c->req->params;
-			$c->stash->{status_msg} = 'Vote saved';
-		}
-		if( $c->req->params->{submit} eq 'Clear vote' ) {
-			delete $c->session->{$formid};
-			$c->stash->{status_msg} = 'Vote cleared';
-		}
-		if( $c->req->params->{submit} eq 'Cast vote' ) {
-			$c->forward('do_public', [ $formid ]);
-		}
+	if( $c->stash->{event}->public_votes_allowed ) 
+	{
+		$c->stash->{candidates} = [ $c->stash->{event}->public_story_candidates ];
+		
+		$c->forward('first_pass') if $c->req->method eq 'POST';
 	}
 	
-	$c->stash->{fillform} = $c->session->{$formid} // undef;
-	
-	push $c->stash->{title}, 'Vote', 'Public';
     $c->stash->{template} = 'vote/public/fic.tt';
 }
 
@@ -50,54 +49,89 @@ sub art :PathPart('vote/public') :Chained('/event/art') :Args(0) {
 	$c->stash->{template} = 'vote/public/art.tt';
 }
 
+sub first_pass :Private {
+	my ( $self, $c ) = @_;
+	
+	if( $c->req->params->{submit} eq 'Save vote' ) 
+	{
+		$c->session->{ $c->stash->{formid} } = $c->req->params;
+		$c->stash->{status_msg} = 'Vote saved';
+	}
+	
+	elsif( $c->req->params->{submit} eq 'Clear vote' ) 
+	{
+		delete $c->session->{ $c->stash->{formid} };
+		$c->stash->{status_msg} = 'Vote cleared';
+	}
+	
+	elsif( $c->req->params->{submit} eq 'Cast vote' ) 
+	{
+		$c->forward('do_public');
+	}
+	
+}
+
 sub do_public :Private {
-	my ( $self, $c, $formid ) = @_;
+	my ( $self, $c ) = @_;
 	
-	return 0 unless $c->stash->{event}->public_votes_allowed;
-	
-	my @candidates = $c->stash->{event}->public_story_candidates( $c->user )->all;
+	my $candidates = $c->stash->{candidates};
 	
 	#The votes are keyed with the id of the story that the votes are cast on
 	my @votes = 
 		grep { defined $c->req->params->{$_} && $c->req->params->{$_} ne 'N/A' }
 		map  { $_->id }
+		grep { $_->user_id != ( $c->user ? $c->user->id : 0 ) }
 		grep { $_->ip ne $c->req->address }
-		@candidates;
+		@$candidates;
 	
 	$c->req->params->{count}   = @votes;
 	$c->req->params->{ip}      = $c->req->address;
+	$c->req->params->{user_id} = $c->user ? $c->user->id : 0;
 	$c->req->params->{captcha} = $c->forward('/captcha_check');
 	
-	my $rs = $c->stash->{event}->vote_records->public->fic;
+	my $rs = $c->stash->{event}->vote_records->public->type( $c->action->name );
 	
 	$c->form(
 		ip       => [ [ 'DBIC_UNIQUE', $rs, 'ip' ] ],
+		user_id  => [ [ 'DBIC_UNIQUE', $rs, 'user_id' ] ],
 		
 		#For whatever reason, FormValidator::Simple doesn't have a >= operator
-		count    => [ [ 'GREATER_THAN', @candidates / 2 - 0.001 ] ],
+		count    => [ [ 'GREATER_THAN', @$candidates / 2 - 0.001 ] ],
 		
 		captcha  => [ [ 'EQUAL_TO', 1 ] ],
 		map { $_ => [ 'NOT_BLANK', 'UINT', [ 'BETWEEN', 0, 10 ] ] } @votes,
 	);
 	
 	if( !$c->form->has_error ) {
-		my $record = $c->stash->{event}->create_related('vote_records', {
+		my %id = (
+			fic => 'story_id',
+			art => 'image_id',
+		);
+		
+		$c->stash->{event}->create_related('vote_records', {
 			user_id => $c->user ? $c->user->get('id') : undef,
 			ip      => $c->req->address,
 			round   => 'public',
 			type    => 'fic',
+			votes   => [ map {
+				{ 
+					$id{ $c->action->name } => $_, 
+					value => $c->form->valid($_) 
+				}
+			} @votes ]
 		});
 		
-		for my $id ( @votes ) {
-			$record->create_related('votes', {
-				story_id => $id,
-				value    => $c->form->valid($id),
-			});
-		}
-		
-		delete $c->session->{$formid};
+		delete $c->session->{ $c->stash->{formid} };
 		$c->stash->{status_msg} = 'Vote successful';
 	}
+}
+
+sub end :Private {
+	my ( $self, $c ) = @_;
+	
+	$c->stash->{fillform} = $c->session->{ $c->stash->{formid} };
+	
+	$c->forward('/end');
 }
 
 =head1 AUTHOR