+++ lib/WriteOff/Schema/ResultSet/Story.pm
@@ -19,16 +19,8 @@ sub with_scores {
 	my $private = $vote_rs->private->search(
 		{ 'private.story_id' => { '=' => { -ident => 'me.id' } } },
 		{
-			select => [{ avg => 'private.value' }],
+			select => [{ sum => 'private.value' }],
 			alias => 'private',
-		}
-	);
-	
-	my $prelim = $vote_rs->prelim->search(
-		{ 'prelim.story_id' => { '=' => { -ident => 'me.id' } } },
-		{
-			select => [{ avg => 'prelim.value' }],
-			alias => 'prelim',
 		}
 	);
 	
@@ -36,9 +28,8 @@ sub with_scores {
 		'+select' => [
 			{ '' => $public->as_query,  -as => 'public_score' },
 			{ '' => $private->as_query, -as => 'private_score' },
-			{ '' => $prelim->as_query,  -as => 'prelim_score' },
 		],
-		'+as' => [ 'public_score', 'private_score', 'prelim_score' ],
+		'+as' => [ 'public_score', 'private_score' ],
 		order_by => [
 			{ -desc => 'private_score' },
 			{ -desc => 'public_score' },
@@ -63,12 +54,65 @@ sub with_stats {
 		$pos_low++ while $pos_low < $n && $this == $storys[$pos_low+1];
 		$this->{__pos_low} = $pos_low;
 		
-		my (@votes, $sum) = $this->votes->get_column('value')->all;
-		$sum += ($_ - $this->public_score) ** 2 for @votes;
-		$this->{__stdev} = sqrt $sum / @votes;
+		my (@votes, $sum) = $this->votes->public->get_column('value')->all;
+		
+		if( @votes ) {
+			$sum += ($_ - $this->public_score) ** 2 for @votes;
+			$this->{__stdev} = sqrt $sum / @votes;
+		}
+		else {
+			$this->{__stdev} = 0;
+		}
 	}
 	
 	return @storys;
+}
+
+sub with_prelim_stats {
+	my $self = shift;
+	
+	my $vote_rs = $self->result_source->schema->resultset('Vote');
+	
+	my $prelim = $vote_rs->prelim->search(
+		{ story_id => { '=' => { -ident => 'me.id' } } },
+		{ alias => 'prelim' }
+	)->get_column('value')->sum_rs;
+	
+	my $record_rs = $self->result_source->schema->resultset('VoteRecord');
+	
+	my $author_vote_count = $record_rs->filled->prelim->search(
+		{
+			user_id  => { '=' => { -ident => 'me.user_id' } },
+			event_id => { '=' => { -ident => 'me.event_id' } },
+		},
+		{
+			group_by => 'record.id',
+			alias => 'record',
+		}
+	)->count_rs;
+	
+	my $author_story_count = $self->search(
+		{ 
+			user_id  => { '=' => { -ident => 'me.user_id' } },
+			event_id => { '=' => { -ident => 'me.event_id' } },
+		},
+		{ alias => 'storys' }
+	)->count_rs;
+	
+	return $self->search_rs(undef, {
+		'+select' => [
+			{ '' => $prelim->as_query, -as => 'prelim_score' },
+			{ '' => $author_vote_count->as_query, -as => 'author_vote_count' },
+			{ '' => $author_story_count->as_query, -as => 'author_story_count' },
+		],
+		'+as' => [ 'prelim_score', 'author_vote_count', 'author_story_count' ]
+	});
+}
+
+sub wordcount {
+	my $self = shift;
+	
+	return $self->get_column('wordcount')->sum;
 }
 
 1;